refactor(mocks): extract notifications module from mockApi (#991 M3)
Move notifications state and notifDTO plus listNotifications/unreadNotificationCount/runsInProgressCount/markNotificationRead into notifications.ts as notificationsApi, spread into index.ts.
